+++ lib/pbio/include/pbio/uartdev.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2018-2020 The Pybricks Authors
+// Copyright (c) 2018-2022 The Pybricks Authors
 
 #ifndef _PBIO_UARTDEV_H_
 #define _PBIO_UARTDEV_H_
@@ -18,6 +18,7 @@ typedef struct {
 #if PBIO_CONFIG_UARTDEV
 
 pbio_error_t pbio_uartdev_get(uint8_t id, pbio_iodev_t **iodev);
+void pbio_uartdev_ready(uint8_t id);
 
 #if !PBIO_CONFIG_UARTDEV_NUM_DEV
 #error Must define PBIO_CONFIG_UARTDEV_NUM_DEV
@@ -30,6 +31,9 @@ extern const pbio_uartdev_platform_data_t pbio_uartdev_platform_data[PBIO_CONFIG
 static inline pbio_error_t pbio_uartdev_get(uint8_t id, pbio_iodev_t **iodev) {
     *iodev = NULL;
     return PBIO_ERROR_NOT_SUPPORTED;
+}
+
+static inline void pbio_uartdev_ready(uint8_t id) {
 }
 
 #endif // PBIO_CONFIG_UARTDEV